+++ src/slop_detector/__init__.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-__version__ = "3.7.8"
+__version__ = "3.7.9"
 __author__ = "Flamehaven Labs"
 __email__ = "info@flamehaven.space"